Renamed txZMQ package to txzmq, to bring the project into PEP8 compliance (see section "Package and Module Names" at http://www.python.org/dev/peps/pep-0008/).
This fixes issue #7.

Detailed changes:
 * updated references to txZMQ in all the modules
 * as well as tests
 * updated example
 * updated README.rst
 * updated setup.py
 * updated Makefile
 * bumped version number to 0.4.0 in setup.py
