--- txZMQ/__init__.py
@@ -1,11 +1,0 @@
-"""
-ZeroMQ integration into Twisted reactor.
-"""
-
-from txZMQ.factory import ZmqFactory
-from txZMQ.connection import ZmqEndpointType, ZmqEndpoint, ZmqConnection
-from txZMQ.pubsub import ZmqPubConnection, ZmqSubConnection
-from txZMQ.xreq_xrep import ZmqXREQConnection
-
-
-__all__ = ['ZmqFactory', 'ZmqEndpointType', 'ZmqEndpoint', 'ZmqConnection', 'ZmqPubConnection', 'ZmqSubConnection', 'ZmqXREQConnection']
