+++ txzmq/connection.py
